+++ lib/daily_report_generator/cli.rb
@@ -1,5 +1,5 @@
 require 'thor'
-require 'daily_report_generator/report_source'
+require 'daily_report_generator/source/events'
 
 module DailyReportGenerator
     class Cli < Thor
@@ -7,19 +7,19 @@ module DailyReportGenerator
 
         desc "github", "generate github daily report"
         def github
-            github_events = DailyReportGenerator::ReportSource.github_events
+            github_events = DailyReportGenerator::Source::Events.github
             github_events.each { |event| puts event.oneline }
         end
 
         desc "google_calendar", "generate google calendar report"
         def google_calendar
-            google_calendar_events = DailyReportGenerator::ReportSource.google_calendar_events
+            google_calendar_events = DailyReportGenerator::Source::Events.google_calendar
             google_calendar_events.each { |event| puts event.oneline }
         end
 
         desc "today", "generate today report"
         def today
-            today_events = DailyReportGenerator::ReportSource.today
+            today_events = DailyReportGenerator::Source::Events.today
             today_events.map { |event| puts event.oneline }
         end
